fix: 修复了评论相关的问题

+++ src/main/java/com/wolfhouse/wolfhouseblog/service/impl/ArticleActionServiceImpl.java
@@ -176,10 +176,14 @@ public class ArticleActionServiceImpl implements ArticleActionService {
                   .doVerify();
 
         Set<Long> ids = getReplyIds(commentId);
-        int i = commentMapper.deleteBatchByIds(ids);
-        if (i != ids.size()) {
-            throw new ServiceException(ArticleConstant.COMMENT_DELETE_FAILED);
+        // 移除回复评论
+        if (!ids.isEmpty()) {
+            int i = commentMapper.deleteBatchByIds(ids);
+            if (i != ids.size()) {
+                throw new ServiceException(ArticleConstant.COMMENT_DELETE_FAILED);
+            }
         }
+
         try {
             return getArticleCommentVosByArticle(dto.getArticleId());
         } catch (Exception e) {
